Completely changed code base
