+++ src/main/java/jenkins/plugins/s3/S3BucketPublisher.java
@@ -1,0 +1,144 @@
+package jenkins.plugins.s3;
+
+import hudson.Extension;
+import hudson.Launcher;
+import hudson.model.BuildListener;
+import hudson.model.Result;
+import hudson.model.AbstractBuild;
+import hudson.model.AbstractProject;
+import hudson.tasks.BuildStepDescriptor;
+import hudson.tasks.BuildStepMonitor;
+import hudson.tasks.Publisher;
+import hudson.tasks.Recorder;
+import hudson.util.CopyOnWriteList;
+
+import java.io.IOException;
+import java.io.PrintStream;
+import java.util.List;
+
+import org.apache.commons.lang.StringUtils;
+import org.kohsuke.stapler.DataBoundConstructor;
+import org.kohsuke.stapler.StaplerRequest;
+
+public final class S3BucketPublisher extends Recorder {
+
+    private String profileName;
+    private List<PublishRule> rules;
+
+    @DataBoundConstructor
+    public S3BucketPublisher(String profileName, List<PublishRule> rules) {
+        super();
+        if (profileName == null) {
+            // defaults to the first one
+            S3Profile[] sites = getDescriptor().getProfiles();
+            if (sites.length > 0)
+                profileName = sites[0].getName();
+        }
+        this.profileName = profileName;
+        this.rules = rules;
+    }
+
+    public String getProfileName() {
+        return this.profileName;
+    }
+
+    public List<PublishRule> getRules() {
+        return this.rules;
+    }
+
+    public BuildStepMonitor getRequiredMonitorService() {
+        return BuildStepMonitor.STEP;
+    }
+
+    public S3Profile getProfile() {
+        S3Profile[] profiles = getDescriptor().getProfiles();
+
+        if (profileName == null && profiles.length > 0)
+            // default
+            return profiles[0];
+
+        for (S3Profile profile : profiles) {
+            if (profile.getName().equals(profileName))
+                return profile;
+        }
+        return null;
+    }
+
+    @Override
+    public boolean perform(AbstractBuild<?, ?> build, Launcher launcher, BuildListener listener)
+            throws InterruptedException, IOException {
+
+        if (build.getResult() == Result.FAILURE) {
+            // build failed. don't post
+            return true;
+        }
+
+        S3Profile profile = getProfile();
+        if (profile == null) {
+            log(listener.getLogger(), "No S3 profile is configured.");
+            build.setResult(Result.UNSTABLE);
+            return true;
+        }
+        log(listener.getLogger(), "Using S3 profile: " + profile.getName());
+        try {
+            for (PublishRule rule : rules) {
+                profile.publish(rule, build.getWorkspace(), listener,
+                        build.getEnvironment(listener));
+            }
+        } catch (IOException e) {
+            e.printStackTrace(listener.error("Failed to upload files"));
+            build.setResult(Result.UNSTABLE);
+        }
+        return true;
+    }
+
+    protected void log(final PrintStream logger, final String message) {
+        logger.println(StringUtils.defaultString(this.getDescriptor().getDisplayName()) + " "
+                + message);
+    }
+
+    @Extension
+    public static final DescriptorImpl DESCRIPTOR = new DescriptorImpl();
+
+    @Override
+    public DescriptorImpl getDescriptor() {
+        return DESCRIPTOR;
+    }
+
+    public static final class DescriptorImpl extends BuildStepDescriptor<Publisher> {
+
+        private final CopyOnWriteList<S3Profile> profiles = new CopyOnWriteList<S3Profile>();
+
+        public DescriptorImpl() {
+            super(S3BucketPublisher.class);
+            load();
+        }
+
+        @Override
+        public String getDisplayName() {
+            return "Publish artifacts to S3 Bucket";
+        }
+
+        @Override
+        public String getHelpFile() {
+            return "/plugin/s3/help.html";
+        }
+
+        @Override
+        public boolean configure(StaplerRequest req, net.sf.json.JSONObject json)
+                throws FormException {
+            profiles.replaceBy(req.bindParametersToList(S3Profile.class, "s3."));
+            save();
+            return true;
+        }
+
+        public S3Profile[] getProfiles() {
+            return profiles.toArray(new S3Profile[0]);
+        }
+
+        @Override
+        public boolean isApplicable(Class<? extends AbstractProject> aClass) {
+            return true;
+        }
+    }
+}
