--- src/main/resources/hudson/plugins/s3/S3BucketPublisher/config.jelly
@@ -1,32 +1,0 @@
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-
-  <j:set var="helpURL" value="/plugin/s3" />
-    <f:entry title="S3 profile">
-      <select class="setting-input" name="s3.name" description="Select configured S3 profile.Check global hudson config for defining access properties for this profile">
-        <j:forEach var="s" items="${descriptor.profiles}">
-          <f:option selected="${s.name==instance.name}">${s.name}</f:option>
-        </j:forEach>
-      </select>
-    </f:entry>
-
-  <f:entry title="Files to upload">
-		
-    <f:repeatable var="e" items="${instance.entries}">
-      <table width="100%">
-        <f:entry title="Source" help="${helpURL}/help-source.html">
-          <input class="setting-input" name="s3.entry.sourceFile"
-            type="text" value="${e.sourceFile}" />
-        </f:entry>
-        <f:entry title="Destination bucket" help="${helpURL}/help-destination.html">
-          <input class="setting-input" name="s3.entry.bucket"
-            type="text" value="${e.bucket}" />
-        </f:entry>
-        <f:entry title="">
-          <div align="right">
-            <f:repeatableDeleteButton />
-          </div>
-        </f:entry>
-      </table>
-    </f:repeatable>
-  </f:entry>
-</j:jelly>
