+++ src/main/resources/jenkins/plugins/s3/S3BucketPublisher/global.jelly
